+++ pages/instructorAssessment/instructorAssessment.sql
@@ -161,6 +161,32 @@ WHERE
 ORDER BY
     aar.number;
 
+-- BLOCK assessment_score_histogram_by_date
+WITH assessment_instances_by_user_and_date AS (
+    SELECT
+        ai.user_id,
+        avg(ai.score_perc) AS score_perc,
+        date_trunc('day', ai.date AT TIME ZONE ci.display_timezone) AS date
+    FROM
+        assessment_instances AS ai
+        JOIN assessments AS a ON (a.id = ai.assessment_id)
+        JOIN course_instances AS ci ON (ci.id = a.course_instance_id)
+    WHERE
+        ai.assessment_id = $assessment_id
+    GROUP BY
+        ai.user_id, date_trunc('day', date AT TIME ZONE ci.display_timezone)
+)
+SELECT
+    ai_by_user_and_date.date,
+    to_char(ai_by_user_and_date.date, 'DD Mon') AS date_formatted,
+    histogram(score_perc, 0, 100, 10)
+FROM
+    assessment_instances_by_user_and_date AS ai_by_user_and_date
+GROUP BY
+    ai_by_user_and_date.date
+ORDER BY
+    ai_by_user_and_date.date;
+
 -- BLOCK assessment_stats
 SELECT * FROM assessment_stats WHERE id = $assessment_id;
 
@@ -179,7 +205,6 @@ SELECT
     hist
 FROM assessment_duration_stats AS ads
 WHERE id = $assessment_id;
-
 
 -- BLOCK assessment_instance_data
 SELECT